\name{weyl-class}
\docType{class}
\alias{weyl-class}
\title{Class \dQuote{weyl}}
\description{The formal S4 class for weyls.}
\section{Objects from the Class}{

Objects \emph{can} be created by calls of the form \code{new("weyl",
  ...)} but this is not encouraged.  Use functions \code{weyl()} or
 \code{as.weyl()} instead.
}
\author{Robin K. S. Hankin}
\keyword{classes}
