\name{x_and_d}
\alias{x_and_d}
\alias{x}
\alias{d}
\docType{data}
\title{Generating elements for the first Weyl algebra}
\description{\loadmathjax

  Variables \code{x} and \code{d} correspond to operator \mjseqn{x} and
  \mjeqn{\partial_x}{omitted}; they are provided for convenience.  These
  elements generate the one-dimensional Weyl algebra.

}
\usage{data(x_and_d)}
\author{Robin K. S. Hankin}
\examples{
d*x-x*d

(1-d*x*d)*(x^2-d^3)

}
\keyword{datasets}
